requirements/dev.txt
-r common.txt
Django
djangorestframework

pytest
pytest-cov